Move VtlRuntimeException
